Added 2nd button; debounced both buttons
Another persistent button is drawn to be used to signal brake squeal.
Both the squeal and the stop/start buttons are now objects of the
'tsbutton' class, which implements buffers to help debounce button
presses and releases and also has functionality for detecting recent
changes to the state of the button (e.g. myButton.wasJustPressed() ).